use std::chrono for main loop timing
